fixed interrupt issues

+++ core/memory/page_frame_allocator.cpp
@@ -202,7 +202,7 @@ void page_frame_allocator::unreserve_page(void* address) {
 }
 
 void page_frame_allocator::unreserve_pages(void* address, uint64_t page_count) {
-	for (int t = 0; t < page_count; t++) {
+	for (uint64_t t = 0; t < page_count; t++) {
 		unreserve_page((void*) ((uint64_t) address + (t * 4096)));
 	}
 }
